# Fixing the SQS Message Size Error
This error occurs because you're trying to send a message to SQS that exceeds the maximum allowed size of 262,144 bytes (256 KB). This is likely happening because you're including the full HTML content of the crawled pages in your results. You need to modify your _send_result method to limit the size of the data being sent to SQS.